requirements.txt
hidapi
flask
pystray
pillow
